Rename function getDockerTimeoutMax
